Conecting Api to Azure Vm

+++ MiApi/Models/Producto.cs
@@ -6,7 +6,7 @@
         public string name { get; set; }
         public string description {  get; set; }   
         public string image {  get; set; }
-        public decimal price { get; set; }
+        public int price { get; set; }
         public int stock { get; set; }
         public string category {  get; set; } 
     }
